rename main script
prediction_Image(Litton-7type-visual-landscape-model).py -> litton7.py
